requirement.txt
click==6.7
pytaglib==1.4.3